Cas de tests basés sur l'exercice de découverte des mocks

+++ src/main/java/co/simplon/PoleEmploi/patrimoine/endpoint/MonumentResource.java
@@ -23,7 +23,7 @@ import co.simplon.PoleEmploi.patrimoine.modele.Monument;
 @RequestScoped
 public class MonumentResource {
 
-	private static int DEFAULT_PAGE_SIZE = 10;
+	protected static int DEFAULT_PAGE_SIZE = 10;
 
 	@Inject
 	private MonumentDao monumentDao;
@@ -64,6 +64,13 @@ public class MonumentResource {
 	@DELETE
 	@Path("{id}")
 	public void deleteMonumentById(@PathParam("id") Long id) {
+		if (id == null) {
+			throw new IllegalArgumentException("L'id doit être renseigné");
+		}
 		monumentDao.deleteMonumentById(id);
+	}
+
+	public void setMonumentDao(MonumentDao monumentDao) {
+		this.monumentDao = monumentDao;
 	}
 }
